Move L4-specific error types and helpers to pkg/l4/utils
