+++ pkg/utils/utils_test.go
@@ -36,7 +36,6 @@ import (
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/cloud-provider-gcp/providers/gce"
 	"k8s.io/ingress-gce/pkg/annotations"
@@ -1069,29 +1068,6 @@ func TestIsQuotaExceededError(t *testing.T) {
 	}
 }
 
-func TestGetErrorType(t *testing.T) {
-	t.Parallel()
-	for _, tc := range []struct {
-		desc    string
-		err     error
-		errType string
-	}{
-		{desc: "nil error", err: nil},
-		{desc: "Forbidden googleapi error", err: &googleapi.Error{Code: http.StatusForbidden}, errType: http.StatusText(http.StatusForbidden)},
-		{desc: "Forbidden googleapi error wrapped", err: fmt.Errorf("Got error: %w", &googleapi.Error{Code: http.StatusForbidden}), errType: http.StatusText(http.StatusForbidden)},
-		{desc: "k8s notFound error", err: k8serrors.NewNotFound(schema.GroupResource{}, ""), errType: "k8s " + string(v1.StatusReasonNotFound)},
-		{desc: "k8s notFound error wrapped", err: fmt.Errorf("Got error: %w", k8serrors.NewNotFound(schema.GroupResource{}, "")), errType: "k8s " + string(v1.StatusReasonNotFound)},
-		{desc: "k8s notFound error embedded with %v", err: fmt.Errorf("Got error: %v", k8serrors.NewNotFound(schema.GroupResource{}, "")), errType: ""},
-		{desc: "unknown error", err: fmt.Errorf("Got unknown error"), errType: ""},
-	} {
-		t.Run(tc.desc, func(t *testing.T) {
-			if errType := utils.GetErrorType(tc.err); errType != tc.errType {
-				t.Errorf("Unexpected errType %q, want %q", errType, tc.errType)
-			}
-		})
-	}
-}
-
 func TestBackendToServicePortID(t *testing.T) {
 	testNS := "test-namespace"
 	for _, tc := range []struct {
@@ -1241,67 +1217,6 @@ func TestMinMaxPortRange(t *testing.T) {
 		portsRange := utils.MinMaxPortRange(tc.svcPorts)
 		if portsRange != tc.expectedRange {
 			t.Errorf("PortRange mismatch %v != %v", tc.expectedRange, portsRange)
-		}
-	}
-}
-
-func TestIsNetworkMismatchGCEError(t *testing.T) {
-	for _, tc := range []struct {
-		err  error
-		want bool
-	}{
-		{
-			err:  fmt.Errorf("The network tier of external IP is STANDARD, that of Address must be the same."),
-			want: true,
-		},
-		{
-			err:  fmt.Errorf("The network tier of external IP is PREMIUM, that of Address must be the same."),
-			want: true,
-		},
-		{
-			err:  fmt.Errorf("The network tier of external IP is , that of Address must be the same."),
-			want: false,
-		},
-		{
-			err:  fmt.Errorf("The network tier of external IP is"),
-			want: false,
-		},
-		{
-			err:  fmt.Errorf("Some dummy string"),
-			want: false,
-		},
-	} {
-		if got := utils.IsNetworkTierMismatchGCEError(tc.err); got != tc.want {
-			t.Errorf("IsNetworkTierMismatchGCEError(%v) = %v, want %v", tc.err, got, tc.want)
-		}
-	}
-}
-
-func TestIsNetworkMismatchError(t *testing.T) {
-	netTierMismatchError := utils.NewNetworkTierErr("forwarding-rule", "premium", "standard")
-	for _, tc := range []struct {
-		description string
-		err         error
-		want        bool
-	}{
-		{
-			description: "Good error is wrapped",
-			err:         fmt.Errorf("err: %w", netTierMismatchError),
-			want:        true,
-		},
-		{
-			description: "Good error is NetworkTierErr type",
-			err:         netTierMismatchError,
-			want:        true,
-		},
-		{
-			description: "Wrong error is not NetworkTierErr type",
-			err:         fmt.Errorf("Wrong error."),
-			want:        false,
-		},
-	} {
-		if got := utils.IsNetworkTierError(tc.err); got != tc.want {
-			t.Errorf("IsNetworkTierError(%v) = %v, want %v", tc.err, got, tc.want)
 		}
 	}
 }
@@ -1560,66 +1475,6 @@ func TestAddIPToLBStatus(t *testing.T) {
 	}
 }
 
-func TestIsUnsupportedFeatureError(t *testing.T) {
-	testCases := []struct {
-		desc                string
-		featureName         string
-		errorMessage        string
-		errorCode           int
-		expectedReturnValue bool
-	}{
-		{
-			desc:                "empty error",
-			featureName:         "randomFeature",
-			errorMessage:        "",
-			errorCode:           200,
-			expectedReturnValue: false,
-		},
-		{
-			desc:                "wrong error code",
-			featureName:         "randomFeature",
-			errorMessage:        "randomFeature is not supported",
-			errorCode:           200,
-			expectedReturnValue: false,
-		},
-		{
-			desc:                "wrong error message",
-			featureName:         "randomFeature",
-			errorMessage:        "randomFeature abc",
-			errorCode:           400,
-			expectedReturnValue: false,
-		},
-		{
-			desc:                "wrong feature name",
-			featureName:         "newFeature",
-			errorMessage:        "randomFeature is not supported",
-			errorCode:           400,
-			expectedReturnValue: false,
-		},
-		{
-			desc:                "feature is not supported",
-			featureName:         "enableStrongAffinity",
-			errorMessage:        "Invalid value for field 'resource.connectionTrackingPolicy.enableStrongAffinity': 'true'. EnableStrongAffinity is not supported., invalid",
-			errorCode:           400,
-			expectedReturnValue: true,
-		},
-	}
-	for _, tc := range testCases {
-		tc := tc
-		t.Run(tc.desc, func(t *testing.T) {
-			t.Parallel()
-
-			err := googleapi.Error{
-				Message: tc.errorMessage,
-				Code:    tc.errorCode,
-			}
-			if utils.IsUnsupportedFeatureError(&err, tc.featureName) != tc.expectedReturnValue {
-				t.Errorf("IsUnsupportedFeatureError returned unexpected result: %v, expectations: %v for error: %v", utils.IsUnsupportedFeatureError(&err, tc.featureName), tc.expectedReturnValue, err)
-			}
-		})
-	}
-}
-
 func TestIsGCEServerError(t *testing.T) {
 	testcases := []struct {
 		desc string
@@ -1863,93 +1718,6 @@ func TestLBBasedOnFinalizer(t *testing.T) {
 
 			if diff := cmp.Diff(tC.want, got); diff != "" {
 				t.Errorf("got != want, diff(-tc.want +got) = %s", diff)
-			}
-		})
-	}
-}
-
-func TestIsConstraintViolationError(t *testing.T) {
-	testCases := []struct {
-		desc string
-		err  error
-		want bool
-	}{
-		{
-			desc: "nil error",
-			err:  nil,
-			want: false,
-		},
-		{
-			desc: "random error",
-			err:  errors.New("random error"),
-			want: false,
-		},
-		{
-			desc: "other google api error",
-			err: &googleapi.Error{
-				Code:    http.StatusBadRequest,
-				Message: "invalid arguments",
-			},
-			want: false,
-		},
-		{
-			desc: "constraint violation google api error",
-			err: &googleapi.Error{
-				Code:    http.StatusPreconditionFailed,
-				Message: "Constraint constraints/compute.restrictLoadBalancerCreationForTypes violated for projects/cf-gcpai-sales-buddy-l-t4. Forwarding Rule projects/cf-gcpai-sales-buddy-l-t4/regions/europe-west4/forwardingRules/k8s2-tcp-mv3ejptg-anthos-identity-servic-gke-oidc-envo-eswdpmuk of type INTERNAL_TCP_UDP is not allowed.",
-			},
-			want: true,
-		},
-		{
-			desc: "status 412 but not constraint violation",
-			err: &googleapi.Error{
-				Code:    http.StatusPreconditionFailed,
-				Message: "precondition failed for some other reason",
-			},
-			want: false,
-		},
-	}
-	for _, tC := range testCases {
-		t.Run(tC.desc, func(t *testing.T) {
-			if got := utils.IsConstraintViolationError(tC.err); got != tC.want {
-				t.Errorf("IsConstraintViolationError(%v) = %v, want %v", tC.err, got, tC.want)
-			}
-		})
-	}
-}
-
-func TestIsIPOutOfRangeError(t *testing.T) {
-	testCases := []struct {
-		desc string
-		err  error
-		want bool
-	}{
-		{
-			desc: "nil error",
-			err:  nil,
-			want: false,
-		},
-		{
-			desc: "other googleapi error",
-			err:  &googleapi.Error{Code: http.StatusBadRequest, Message: "Some other error"},
-			want: false,
-		},
-		{
-			desc: "not a googleapi error",
-			err:  fmt.Errorf("Requested internal IP address is outside the network/subnetwork range"),
-			want: false,
-		},
-		{
-			desc: "correct googleapi error",
-			err:  &googleapi.Error{Code: http.StatusBadRequest, Message: "Invalid value for field 'resource.ipAddress': '10.24.120.53'. Requested internal IP address is outside the network/subnetwork range., invalid"},
-			want: true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.desc, func(t *testing.T) {
-			if got := utils.IsIPOutOfRangeError(tc.err); got != tc.want {
-				t.Errorf("IsIPOutOfRangeError(%v) = %v, want %v", tc.err, got, tc.want)
 			}
 		})
 	}
